Fixing compilation errors that were intrudoced with the new templated math functions

+++ Source/Urho3D/Core/Profiler.cpp
@@ -106,7 +106,7 @@ void Profiler::PrintData(ProfilerBlock* block, String& output, unsigned depth, u
     char line[LINE_MAX_LENGTH];
     char indentedName[LINE_MAX_LENGTH];
 
-    unsigned intervalFrames = (unsigned)Max(intervalFrames_, 1);
+    unsigned intervalFrames = Max(intervalFrames_, unsigned(1));
 
     if (depth >= maxDepth)
         return;
@@ -129,7 +129,7 @@ void Profiler::PrintData(ProfilerBlock* block, String& output, unsigned depth, u
                 float frame = block->intervalTime_ / intervalFrames / 1000.0f;
                 float all = block->intervalTime_ / 1000.0f;
 
-                sprintf(line, "%s %5u %8.3f %8.3f %8.3f %9.3f\n", indentedName, Min(block->intervalCount_, 99999),
+                sprintf(line, "%s %5u %8.3f %8.3f %8.3f %9.3f\n", indentedName, Min(block->intervalCount_, unsigned(99999)),
                     avg, max, frame, all);
             }
             else
@@ -142,8 +142,8 @@ void Profiler::PrintData(ProfilerBlock* block, String& output, unsigned depth, u
                 float totalMax = block->totalMaxTime_ / 1000.0f;
                 float totalAll = block->totalTime_ / 1000.0f;
 
-                sprintf(line, "%s %5u %8.3f %8.3f %9.3f  %7u %9.3f %9.3f %11.3f\n", indentedName, Min(block->frameCount_, 99999),
-                    avg, max, all, Min(block->totalCount_, 99999), totalAvg, totalMax, totalAll);
+                sprintf(line, "%s %5u %8.3f %8.3f %9.3f  %7u %9.3f %9.3f %11.3f\n", indentedName, Min(block->frameCount_, unsigned(99999)),
+                    avg, max, all, Min(block->totalCount_, unsigned(99999)), totalAvg, totalMax, totalAll);
             }
 
             output += String(line);
